+++ src/redux/reducer.js
@@ -23,7 +23,7 @@ export const INITIAL_STATE = {
 function mergeNativeAndThirdPartyReducers (reducers, normaliser) {
   const baseReducer = {
     [ActionTypes.RequestAck]: [acknowledgeReducer],
-    [ActionTypes.RequestComplete]: [completeReducer(normaliser)],
+    [ActionTypes.RequestComplete]: [createCompleteReducer(normaliser)],
     [ActionTypes.RequestFail]: [failReducer],
     [ActionTypes.RewindQueryCounter]: [rewindReducer],
     [ActionTypes.IncrementNextQueryId]: [incrementNextQueryIdReducer]
@@ -43,10 +43,16 @@ function mergeNativeAndThirdPartyReducers (reducers, normaliser) {
 
   // Produce a single function for each action type
   for (const actionType in reducersByActionType) {
+    if (!reducersByActionType.hasOwnProperty(actionType)) {
+      continue
+    }
+    
     if (reducersByActionType[actionType].length > 1) {
       // Call each reducer function in succession, passing the state returned from each to the next
-      reducer[actionType] = (state, action) => {
-        return reducersByActionType[actionType].reduce((state, fn) => fn(state, action), state)
+      reducer[actionType] = (state, action, kasiaPluginUtils) => {
+        return reducersByActionType[actionType].reduce((state, fn) => {
+          return fn(state, action, kasiaPluginUtils)
+        }, state)
       }
     } else {
       // Take the first and only function as the whole reducer
@@ -75,23 +81,26 @@ export function acknowledgeReducer (state, {request}) {
 
 // COMPLETE
 // Place entity on the store; update query record if for component (has an id)
-export function completeReducer (normalise) {
+export function createCompleteReducer (normalise) {
   return (state, {request}) => {
     const {__keyEntitiesBy: idAttribute} = state
     const query = state.queries[request.id]
     const newState = merge({}, state)
 
     if (request.id !== PreloadQueryId && !query) {
-      console.log('[kasia] attempt to complete non-existent query. Ignoring. Request:', {request})
+      console.log('[kasia] ignoring attempt to complete non-existent request:', request)
       return newState
     }
 
-    newState.entities = merge(
-      newState.entities,
-      normalise(request.result, {idAttribute})
-    )
+    const mergeInNormalised = () => {
+      newState.entities = merge(
+        newState.entities,
+        normalise(request.result, {idAttribute})
+      )
+    }
 
     if (request.id === PreloadQueryId) {
+      mergeInNormalised()
       return newState
     }
 
@@ -106,6 +115,7 @@ export function completeReducer (normalise) {
       newQuery.result = request.result
       newQuery.paging = null
     } else {
+      mergeInNormalised()
       newQuery.entities = pickEntityIds(request.result)
       newQuery.paging = request.result._paging || {}
     }
@@ -164,7 +174,9 @@ export default function createReducer ({ keyEntitiesBy, reducers }) {
       const [ actionNamespace ] = action.type.split('/')
 
       if (actionNamespace === 'kasia' && action.type in reducer) {
-        return reducer[action.type](state, action)
+        return reducer[action.type](state, action, {
+          normaliseEntities: normaliser
+        })
       }
 
       return state