+++ src/util/query-builder.js
@@ -1,12 +1,5 @@
 import contentTypesManager from './content-types-manager'
 
-/** Fetch data for a single post via the `wpapi` instance. */
-function queryFn (wpapi, contentTypeMethodName, idTypeMethodName, id) {
-  const contentTypeApi = wpapi[contentTypeMethodName]()
-  const query = contentTypeApi[idTypeMethodName](id)
-  return query.embed().get()
-}
-
 /**
  * Create a function that dynamically calls the necessary wpapi
  * methods that will fetch data for the given content type item
@@ -23,8 +16,12 @@ function queryFn (wpapi, contentTypeMethodName, idTypeMethodName, id) {
  * @returns {Function} A function to make a request to the WP-API
  */
 export function deriveQueryFunction (contentTypeMethodName, identifier) {
-  const idMethodName = typeof identifier === 'string' ? 'slug' : 'id'
-  return (wpapi) => queryFn(wpapi, contentTypeMethodName, idMethodName, identifier)
+  return (wpapi) => {
+    const idTypeMethodName = typeof identifier === 'string' ? 'slug' : 'id'
+    const contentTypeApi = wpapi[contentTypeMethodName]()
+    const query = contentTypeApi[idTypeMethodName](identifier)
+    return query.embed().get()
+  }
 }
 
 /** Given an `action` produce a function that will query the WP-API. */