+++ test/__mocks__/components/BuiltInContentType.js
@@ -17,10 +17,6 @@ export class target extends Component {
       return <div>Loading...</div>
     }
 
-    if (!post) {
-      console.log(query, post)
-    }
-
     return <div>{post.title.rendered}</div>
   }
 }