+++ test/redux/sagas.js
@@ -14,19 +14,28 @@ import { ActionTypes, ContentTypes } from '../../src/constants'
 jest.disableAutomock()
 
 describe('redux/sagas', () => {
+  const QUERY_ID_SELECTOR_RESULT = 0
+  const REQUEST_RESULT = 'mockResult'
+
   describe('createPostRequest', () => {
-    const action = createPostRequest(ContentTypes.Post, 16)
+    const action = createPostRequest({
+      contentType: ContentTypes.Post,
+      identifier: 16
+    })
     const generator = fetch(action)
 
     it('yields a select _getCurrentQueryId', () => {
       const actual = generator.next().value
       const expected = select(_getCurrentQueryId)
-      expect(actual.SELECT.selector).toEqual(expected.SELECT.selector)
+      expect(actual).toEqual(expected)
     })
 
     it('yields a put acknowledgeRequest', () => {
-      const actual = generator.next().value
-      const expected = put({ ...action, type: ActionTypes.RequestAck })
+      const actual = generator.next(QUERY_ID_SELECTOR_RESULT).value
+      const expected = put({
+        ...action,
+        type: ActionTypes.RequestAck
+      })
       expect(actual).toEqual(expected)
     })
 
@@ -39,11 +48,13 @@ describe('redux/sagas', () => {
     })
 
     it('yields a put completeRequest', () => {
-      const actual = generator.next('mockResult').value
+      const actual = generator.next(REQUEST_RESULT).value
       const expected = put({
         type: ActionTypes.RequestComplete,
         request: {
-          result: 'mockResult'
+          id: 0,
+          fromCache: false,
+          result: REQUEST_RESULT
         }
       })
       expect(actual).toEqual(expected)
@@ -52,7 +63,7 @@ describe('redux/sagas', () => {
 
   describe('createQueryRequest', () => {
     const queryFn = () => {}
-    const action = createQueryRequest(queryFn)
+    const action = createQueryRequest({ queryFn })
     const generator = fetch(action)
 
     it('yields a select _getCurrentQueryId', () => {
@@ -62,7 +73,7 @@ describe('redux/sagas', () => {
     })
 
     it('yields a put acknowledgeRequest', () => {
-      const actual = generator.next().value
+      const actual = generator.next(QUERY_ID_SELECTOR_RESULT).value
       const expected = put({
         ...action,
         type: ActionTypes.RequestAck
@@ -77,11 +88,13 @@ describe('redux/sagas', () => {
     })
 
     it('puts a put completeRequest with result', () => {
-      const actual = generator.next('mockResult').value
+      const actual = generator.next(REQUEST_RESULT).value
       const expected = put({
         type: ActionTypes.RequestComplete,
         request: {
-          result: 'mockResult'
+          id: 0,
+          fromCache: false,
+          result: REQUEST_RESULT
         }
       })
       expect(actual).toEqual(expected)